+++ libnavigation-base/src/test/java/com/mapbox/navigation/base/internal/CopilotOptionsTest.kt
@@ -1,8 +1,11 @@
 package com.mapbox.navigation.base.internal
 
+import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
+import com.mapbox.navigation.base.options.CopilotOptions
 import com.mapbox.navigation.testing.BuilderTest
 import org.junit.Test
 
+@OptIn(ExperimentalPreviewMapboxNavigationAPI::class)
 class CopilotOptionsTest : BuilderTest<CopilotOptions, CopilotOptions.Builder>() {
     override fun getImplementationClass() = CopilotOptions::class